+++ src/main/java/persistence/sql/ddl/metadata/ColumnMetadata.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.metadata;
 
 import jakarta.persistence.Transient;
