+++ src/main/java/persistence/sql/ddl/metadata/ColumnOption.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.metadata;
 
 public enum ColumnOption {
     NOT_NULL("not null"),
