+++ src/main/java/persistence/sql/ddl/metadata/EntityMetadata.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.metadata;
 
 import java.util.List;
