+++ src/test/java/persistence/sql/ddl/metadata/ColumnMetadataTest.java
@@ -1,12 +1,10 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.metadata;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.sql.ddl.fixture.IncludeId;
 import persistence.sql.ddl.fixture.NotIncludeId;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
